==> 1.1_ETL_P1_games.ipynb <==
## ETL GAMES

# Se cargan las librerias necesarias
import pandas as pd 
import numpy as np 
import gzip
import json
from pandas import json_normalize
import ast

Extraccion y exploracion del dataset games

# Lectura de datos cunado el archivo esta descomprimido
#games = pd.read_json('output_steam_games.json', lines=True)
#games

# Ruta del archivo comprimido 
ruta_archivo_comprimido = 'steam_games.json.gz'

# Nombre del archivo JSON dentro del archivo comprimido
nombre_archivo_json = 'output_steam_games.json'

# Abre el archivo comprimido en modo lectura
with gzip.open(ruta_archivo_comprimido, 'rb') as archivo_gzip:
    # Lee el contenido del archivo Gzip y carga los datos en un DataFrame de Pandas
    games = pd.read_json(archivo_gzip, lines=True)

games

Se borran las columnas que tiene todos sus valores nulos

games = games.dropna(how='all').reset_index(drop=True)

games.info()

games.head()

### Estructura de la informacion
* __publisher__: empresa quer publica el juego.
* __genres__: genero del juego. Pueden tener varios
* __app_name__: Nombre del juego.
* __title__: titulo del juego.
* __url__: url del juego.
* __release_date__: fecha de lanzamiento del juego formato YY-MM-DD.
* __tags__: etiqueta del juego. Puede tener varias.
* __reviews_url__: url donde se encuentra el review del juego.
* __specs__: especificaciones de cada juego. Puede tener varias.
* __price__: Precio del juego
* __early_access__:  el acceso temprano 
* __id__: identificador único del juego.
* __developer__: desarrollador del juego.


Se revisa si hay duplicados

duplicados = games[games.duplicated(subset='id', keep=False)]
duplicados

Hay 2 filas que parecen iguales pero faltan datos index 74 y 30961

Se hace una busqueda por title para ver si hay otra fila con mas informacion

games[games['title'] == 'Batman: Arkham City - Game of the Year Edition']

Se puede observar que la fila si esta duplicada y se procede a borrar estos datos

index_drop = [74, 14573, 30961]
games = games.drop(index_drop)
games = games.drop_duplicates(subset='id', keep='first')
duplicados = games.duplicated(subset='id')
duplicados =duplicados.sum()
print("La cantidad de filas duplicadas es:", duplicados)


Se extrae el año de lanzamiento del juego y se elimina la columna release_date

games['rel_year'] = games['release_date'].str.extract(r'(\d{4})')
games.loc[games['rel_year'].isnull(), 'rel_year'] = 'No_disponible'
games.drop('release_date', inplace=True, axis=1)

games.info()

En la columna price hay texto, donde dice sobre promociones o que es gratis. Se decide volver el texto en valor 0 

games['price'] = pd.to_numeric(games['price'], errors='coerce').fillna(0)
games['price'].dtype

games['price'].value_counts()

Las siguientes columnas 'publisher', 'app_name', 'title', 'developer' se decide llenar con No_disponible donde hay valores faltantes o nulos

c_l = ['publisher', 'app_name', 'title', 'developer']
games[c_l] = games[c_l].fillna('No_disponible')


games.head()

En la columna genres solo debe quedar un valor para que sea mas facil el manejo despues

games = games.explode('genres')
games = games.dropna(subset=['genres'])
games.head()

Se borran las columnas 'tags', 'specs', 'url' y 'reviews_url' pues no son necesarias para el analisis posterior

Se revisa si hay valores nulos

games = games.drop(['tags', 'specs', 'url', 'reviews_url'], axis=1)
games.info()

Se guarda el dataset limpio

games.to_csv('games_clean.csv', index=False, encoding='utf-8')
print(f'Se guardó el archivo')
