==> 1.3_ETL_P1_review.ipynb <==
## ETL USER


# Se cargan las librerias necesarias
import pandas as pd 
import numpy as np 
import json
from pandas import json_normalize
import ast
import gzip


Extraccion y exploracion del dataset reviews

# CArga del archivo cuando esta descomprimido
# Se leen los datos, se guardan en un dataframe y se visualiza la estructura que tiene
#reviews_ = []
#with open('australian_user_reviews.json', 'r', encoding='utf-8') as r:
#    for line in r:
#        reviews_.append(ast.literal_eval(line))

#reviews = pd.DataFrame(reviews_)
#reviews.head()


reviews_ = []

# Ruta al archivo comprimido que contiene los datos en formato Gzip
ruta_archivo_comprimido = 'user_reviews.json.gz'

# Abre el archivo comprimido en modo lectura
with gzip.open(ruta_archivo_comprimido, 'rt', encoding='utf-8') as archivo_gzip:
    # Lee el contenido del archivo Gzip línea por línea
    for line in archivo_gzip:
        # Usa ast.literal_eval para interpretar cada línea como un objeto Python
        review = ast.literal_eval(line)
        reviews_.append(review)

# Crea un DataFrame de Pandas con los datos
reviews = pd.DataFrame(reviews_)
reviews.head()

Se verifica el tipo de datos y la cantidad de datos de cada columna del dataframe

print(reviews.info())

Se busca si hay filas duplicadas

duplicados = reviews.duplicated(subset='user_id', keep=False)

filas_duplicadas = reviews[duplicados]

print("La cantidad de filas duplicadas es:", len(filas_duplicadas)) 
filas_duplicadas.head()  

En el dataframe reviews  se revisa la estructura que tiene la columna reviews para poder tomar una decision con las filas duplicadas

reviews['reviews'][0]

### Estructura de la informacion
* __user_id__: identificador único para el usuario.
* __user_url__: url del perfil del usuario en streamcommunity.
* __reviews__: contiene una lista de diccionarios. Para cada usuario se tiene uno o mas diccionarios. Cada diccionario contiene:
    * __funny__: indica si alguien puso emoticón de gracioso al review.
    * __posted__: fecha de posteo del review en formato Posted April 21, 2011.
    * __last_edited__: fecha de la última edición.
    * __item_id__: identificador único del juego.
    * __helpful__: estadística donde otros usuarios indican si fue útil la información.
    * __recommend__: booleano que indica si el usuario recomienda o no el juego.
    * __review__: comentarios sobre el juego.

Se escojen 2 user_id al azar para comprobar si realmente son filas duplicadas

# Se buscan dos user_id diferentes al azar
user_ids = filas_duplicadas['user_id'].sample(n=2, replace=False, random_state=42).tolist()

# Se Filtra el DataFrame para obtener las filas de los dos user_id seleccionados
filas = filas_duplicadas[filas_duplicadas['user_id'].isin(user_ids)]

# Imprimir las filas seleccionadas
print("Filas seleccionadas para los user_id:", user_ids)
filas.head()

# Verificar e imprimir el primer review de la lista de diccionarios
for user_id, group in filas.groupby('user_id'):
    print(f"\nPara el user_id {user_id}:")
    for index, row in group.iterrows():
        first_review = row['reviews'][0]['review']
        print(f"Primer review: {first_review}")


Mediante esta prueba se puede verificar que no solo estan duplicados los user_id sino tambien la demas informacion.

Teniendo en cuenta lo anterior se procede a borrar los duplicados.

reviews = reviews.drop_duplicates(subset='user_id', keep='first')
duplicados = reviews.duplicated(subset='user_id')
duplicados =duplicados.sum()
print("La cantidad de filas duplicadas es:", duplicados)

Desanidar la columna reviews del dataframe

# Desanidar la columna 'reviews'
reviews_des = pd.concat([reviews.drop(['reviews'], axis=1), 
                        json_normalize(reviews['reviews'])], axis=1)
reviews_des.head()

Las nuevas columnas que son ahora numeros se deben volver filas

reviews_des = pd.melt(reviews_des, id_vars=['user_id', 'user_url'], value_vars=list(range(9)), value_name='reviews')
reviews_des.head()


Se toma un user_id y se observa como quedo

reviews_des[reviews_des['user_id'] == 'evcentric']

Se comprueba la forma de los datos y se procede a borrar las filas que tengan None

reviews_des = reviews_des.dropna()
reviews_des[reviews_des['user_id'] == 'evcentric']

Ahora del diccionario que esta en la columna reviews, cada clave se vuleve una columna nueva dataframe

# Se toman los diciconarios de la columna reviews y se vuelven series de pandas y despues se unen con dataframe 
expanded_reviews = reviews_des['reviews'].apply(pd.Series)
expanded_reviews = expanded_reviews.add_prefix('rev_')
reviews_des = pd.concat([reviews_des, expanded_reviews], axis=1).drop('reviews', axis=1)
reviews_des.head()

Se puede observar campos vacios, se procede a cambiarlos por None y se verifica si quedan aun campos vacios

reviews_des.replace('', None, inplace=True)
(reviews_des == '').sum()

reviews_des.info()

Las columnas rev_funny y rev_last_edited tienen muy pocos datos, se toma la decision de borrarlas

reviews_des = reviews_des.drop(columns=['rev_funny', 'rev_last_edited'])

Se extrae de la columna rev_posted el año en que fue hecho

# Se extrae el año y se forma una nueva columna y se le da manejo a las filas que no se puede leer el año
reviews_des['año'] = reviews_des['rev_posted'].str.extract(r'(\d{4})')
reviews_des.loc[reviews_des['año'].isnull(), 'año'] = 'invalido'
reviews_des['año']

reviews_des[reviews_des['año'] == 'invalido']

Se elimina la columna rev_posted pues ya no tiene mas informacion relevante para el trabajo a realizar y ademas se borran los valores nulos de la columna rev_review	
pues no es mas del 5% de  los datos

reviews_des = reviews_des.drop('rev_posted', axis=1)
reviews_des = reviews_des.dropna(subset=['rev_review'])
reviews_des.info()

Se guarda el dataset limpio

reviews_des.to_csv('review_clean.csv', index=False, encoding='utf-8')
print(f'Se guardó el archivo')
