==> 3_desarrollo_funcion_api.ipynb <==
# Se cargan las librerias necesarias
import pandas as pd 
import numpy as np 


Se cargan los datasets para resolver cada funcion

###  Funcion PlayTimeGenre

Para devolver el año con mas horas jugadas en un determinado genero

playtime = pd.read_csv('playtime.csv')
playtime.head()

Funcion para la API

def PlayTimeGenre(genre):
    genre = genre.lower()
    # Filtrar el DataFrame por el género proporcionado
    filtered_data = playtime[playtime['genres'].str.lower() == genre]

    if filtered_data.empty:
        return {"No hay datos para el género."}
    
    # Agrupar por año y sumar las horas de juego
    grouped = filtered_data.groupby('rel_year')['playtime_forever'].sum().reset_index()
    
    # Encontrar el año con la máxima cantidad de horas de juego
    año_mas_jugado = grouped.loc[grouped['playtime_forever'].idxmax(), 'rel_year']
    
    return print("Año de lanzamiento con más horas jugadas para Género :", genre, int(año_mas_jugado))

PlayTimeGenre('acTiOn')
# http://127.0.0.1:8000/PlayTimeGenre/?genre=action

### Funcion UserForGenre
Para devolver el usuario que acumula mas horas jugadas para un determinado genero y una lista de la acumulacion de horas jugadas por año

userfor = pd.read_csv('userfor.csv')
userfor.head()

Funcion para la API

def UserForGenre(genre):
    genre = genre.lower()
    # Filtrar el DataFrame por el género dado
    genero = userfor[userfor['genres'].str.lower() == genre]
    
    if genero.empty:
        #print(f"No hay datos para el género '{genre}'.")
        return  {"No hay datos para el género."}
    
    # Encontrar el usuario que más horas ha jugado ese género
    max_user = genero.loc[genero['playtime_forever'].idxmax()]['user_id']
    # Se filtra el DataFrame para obtener solo las filas correspondientes al usuario con más horas jugadas en ese género
    usuario = genero[genero['user_id'] == max_user]

    
    # Calcular la acumulación de horas jugadas por año para ese género y usuario               
    agrupar = usuario.groupby('rel_year')['playtime_forever'].sum().reset_index()
    
    # Se vuelve genre_yearly_playtime en un diccionario para mostrar los resultados
    agrupar.columns = ['Year', 'Time']
    agrupar_dict = agrupar.to_dict(orient='records')
    result = {"Usuario con más horas jugadas para Género " + genre: max_user,
        "Horas jugadas": agrupar_dict}
    
    return result

# Llamada a la función con el género deseado (por ejemplo, 'Action')
UserForGenre('AcTion')
# http://127.0.0.1:8000/UserForGenre/?genre=action

### Funcion UsersRecommend
Para devolver el top 3 de juegos mas recomendados por usuarios para el año dado <br>

user_recom = pd.read_csv('user_recom.csv')
user_recom.head()

Funcion para la API

def UsersRecommend(año):
    # Filtrar el DataFrame por el año proporcionado
    filtro = user_recom[user_recom['año'] == año]

    if filtro.empty:
        return  {"No hay datos para el año."}
    
    # Obtener los nombres de app_name en el mismo orden que aparecen en el DataFrame original
    nombres_app = filtro['app_name'].tolist()
    resultado = {f'Puesto {i+1}': app for i, app in enumerate(nombres_app)}
    
    return resultado

UsersRecommend(2011)
# http://127.0.0.1:8000/UsersRecommend/?año=2011

### Funcion UsersWorstDeveloper 
Para devolver el top 3 de las desarrolladoras con juegos menos recomendados por usuarios para el año dado

user_worst = pd.read_csv('user_worst.csv')
user_worst.head()

Funcion para la API

def UsersWorstDeveloper(año):
    # Filtrar el DataFrame por el año proporcionado
    filtro = user_worst[user_worst['año'] == año]

    if filtro.empty:
        return  {"No hay datos para el año."}

    # Obtener los nombres de app_name en el mismo orden que aparecen en el DataFrame original
    developer = filtro['developer'].tolist()
    resultado = {f'Puesto {i+1}': app for i, app in enumerate(developer)}
    
    return resultado

UsersWorstDeveloper(2012)
# http://127.0.0.1:8000/UsersWorstDeveloper/?año=2012

### Funcion sentiment_analysis 
Segun la empresa desarrolladora se devuelve un diccionario donde la clave es la desarrolladora y una lista con la cantidad total de registros<br>
positivos, negativos y neutrales. <br>

sa = pd.read_csv('sa.csv')
sa.head()

def sentiment_analysis(developer):
    developer = developer.lower()
    filtro = sa.loc[sa['developer'].str.lower() == developer]

    if filtro.empty:
        return  {"No hay datos para el desarrollador."}
    
    resultado = {
        developer: {
            'Negative': filtro.iloc[0, 1], 
            'Neutral': filtro.iloc[0, 2],   
            'Positive': filtro.iloc[0, 3]  
                         }
                }
    return resultado



# Llamada a la función
sentiment_analysis('Gaijin Games')
# http://127.0.0.1:8000/sentiment_analysis/?developer=Gaijin Games
