==> 4.1_Dataset_sistema_recomendacion.ipynb <==
### Dataset para el sistema de Recomendacion

# Se cargan las librerias necesarias
import pandas as pd 
import numpy as np 

# Se cargan los datos
reviews = pd.read_csv('reviews_sa.csv', encoding='utf-8')
items = pd.read_parquet('item_clean.parquet', engine='auto')
items['item_id'] = items['item_id'].astype(int)
games = pd.read_csv('games_clean.csv', encoding='utf-8')

Para este dataset se necesita un Rating o valoracion de cada item, para nuestro caso en el dataframe reviews se tienen 2, el sentiment_analysis y rev_recommend. <br>
Se diseña un sistema de valoracion teniendo en cuenta lo anterior de la siguiente manera: <br> 
Cuando: <br>
sentiment_analysis = 0  y rev_recommend = 0  entonces Rating = 1  <br>
sentiment_analysis = 0  y rev_recommend = 1  entonces Rating = 1  <br>
sentiment_analysis = 1  y rev_recommend = 0  entonces Rating = 2  <br>
sentiment_analysis = 1  y rev_recommend = 1  entonces Rating = 3  <br>
sentiment_analysis = 2  y rev_recommend = 0  entonces Rating = 4  <br>
sentiment_analysis = 2  y rev_recommend = 1  entonces Rating = 5  <br>


Volviendo lo anterior codigo y aplicandolo a games tenemos:

# Se definen las condiciones y los valores a asignar
conditions = [
    (reviews['sentiment_analysis'] == 0) & (reviews['rev_recommend'] == 0),
    (reviews['sentiment_analysis'] == 0) & (reviews['rev_recommend'] == 1),
    (reviews['sentiment_analysis'] == 1) & (reviews['rev_recommend'] == 0),
    (reviews['sentiment_analysis'] == 1) & (reviews['rev_recommend'] == 1),
    (reviews['sentiment_analysis'] == 2) & (reviews['rev_recommend'] == 0),
    (reviews['sentiment_analysis'] == 2) & (reviews['rev_recommend'] == 1)
]

# Valores para cada condición
values = [1, 1, 2, 3, 4, 5]  

# Se aplican las condiciones y se asignan los valores usando np.select()
reviews['Rating'] = np.select(conditions, values, default=None)
reviews.head()

Se hacen  tablas auxiliares  para tomar los datos necesarios para el sistema de recomendacion

reviews_aux =  reviews[['rev_item_id', 'user_id', 'Rating']]
reviews_aux = reviews_aux.rename(columns={'rev_item_id':'item_id'})

# Se usa la items para traer los nombres de los juegos, se usa esta tabla porque tiene mayor numero de juegos 
items_aux = items[['item_id', 'item_name']]
items_aux = items_aux.drop_duplicates(subset=['item_id'])

Se hace la union de las tablas

training = reviews_aux.merge(items_aux, on='item_id')
training = training.drop_duplicates()
training.head()

training.item_id.nunique()

training.to_parquet('training.parquet')
print(f'Se guardó el archivo')
