==> 5_sistema_recomendacion.ipynb <==
### SISTEMA DE RECOMENDACION

# Se cargan las librerias necesarias
import pandas as pd 
import numpy as np 
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import LabelEncoder
import scipy as sp
import pickle
import gzip

# Cargar los datos
training = pd.read_parquet('training.parquet')
training

Se hace una copia de la tabla para trabajar con ella

training_1 = training.drop(columns=['item_name']).copy()
training_1

### Funcion  recomendacion de juegos

Se normalizan los datos para mejorar los resultados

# Se normaliza la columna Rating para mejorar los resultados
scaler = MinMaxScaler()
training_1['Norm_Rating'] = scaler.fit_transform(training_1[['Rating']])

# Elimino la columna Rating
training_1.drop(columns=['Rating'], inplace=True)

Se pivotea la tabla para dejar los datos de la forma adecuada para que puedan ser tratados por la funcion cosine_similarity

# Se crea la matriz usuario-juego para poder usar cosine_similarity
matriz_1 = training_1.pivot_table(index=['user_id'], columns=['item_id'], values='Norm_Rating').fillna(0)
matriz_1

Se convierte la matriz a fortmato matriz dispersa (sparse matrix) para redfucir la memoria utilizada y mejorar la eficencia

recomendacion_sparse = sp.sparse.csr_matrix(matriz_1.values)

Se instancia cosine_similarity en los datos, teniendo en cuenta que para item es la traspuesta de la matriz que se tiene hasta el momento

item_cos = cosine_similarity(recomendacion_sparse.T)

Se organiza y estructura los resultados en un dataframe

df_item_cos = pd.DataFrame(item_cos, index = matriz_1.columns, columns = matriz_1.columns)

Se hace la funcion para el sistema de recomendacion para juegos similares

def recomendacion_juego(game: str):
    recomendacion = []
    count = 1
    if game not in df_item_cos.columns:
        return {"Mensaje": f"El juego con el ID {game} no está en el listado de juegos."}

    print('Juegos similares a {} :\n'.format(game))
    juegos_similares = df_item_cos.sort_values(by=game, ascending=False).index[1:6]
    
    for item in juegos_similares:
        item_name = training[training['item_id'] == item]['item_name'].values[0]
        recomendacion.append({'No.': count, 'Juego': item_name, 'Id del juego': item})
        count += 1
    
    return {"Juegos similares a {}".format(game): recomendacion}

Prueba del codigo

recomendacion_juego(10)
# http://127.0.0.1:8000/recomendacion_juego/?game=10

df_item_cos.to_csv('df_item_cos.csv', index=True, header=True)
training.to_csv('training.csv', index=False)
print(f'Se guardó el archivo')

df_item_cos.shape

### Funcion para recomendar juegos a un usuario

Ahora se hace la funcion para recomendar a un usuario juegos similares a los que juega

# Se filtran los valores para tener una matriz mas pequeña y se pueda lograr el deploy en render
# Se buscan 3000 usuarios al azar
np.random.seed(42)
user_ids_aleatorios = np.random.choice(training_1['user_id'].unique(), size=3000, replace=False)
training_1 = training_1[training_1['user_id'].isin(user_ids_aleatorios)]
training_1

# Se convierte  el user_id a valores numericos usando LabelEncoder
label_encoder = LabelEncoder()
#training_1['user_id_encoded'] = label_encoder.fit_transform(training_1['user_id'])
training_1.loc[:, 'user_id_encoded'] = label_encoder.fit_transform(training_1['user_id'])

# Se crea la matriz usuario-juego para poder usar cosine_similarity
matriz_2 = training_1.pivot_table(index='user_id_encoded', columns='item_id', values='Norm_Rating', fill_value=0)
matriz_2

# Se convierte la matriz a una matriz dispersa (sparse matrix)
matriz_2_sparse = sp.sparse.csr_matrix(matriz_2.values)

# Se instancia el coseno de similaridad 
cosine_sim = cosine_similarity(matriz_2_sparse, matriz_2_sparse)
cosine_sim

Se hace la funcion para recomendar juegos a un usuario especifico

def recomendacion_usuario(user_id):
    n=5
    user_id = user_id.lower()

    # Verificar si el user_id está en el DataFrame
    if user_id not in training_1['user_id'].str.lower().unique():
        return {"Mensaje": f"El usuario con el ID {user_id} no está registrado en los datos."}
    
    # Encuentra el índice del usuario en el DataFrame
    user_index = training_1[training_1['user_id'].str.lower() == user_id]['user_id_encoded'].iloc[0]
    
    # Obtener juegos similares ordenados por similitud de coseno
    juegos_similares = list(enumerate(cosine_sim[user_index]))
    juegos_similares = sorted(juegos_similares, key=lambda x: x[1], reverse=True)
    
    # Obtener los juegos ya jugados por el usuario
    juegos_jugados = set(training_1[training_1['user_id'].str.lower() == user_id]['item_id'].tolist())
    
    # Filtrar juegos similares que ya han sido jugados y no están en la lista
    top_juegos_similares = []
    for idx, _ in juegos_similares:
        if len(top_juegos_similares) >= n:
            break
        game_id = training_1.iloc[idx]['item_id']
        if game_id not in juegos_jugados and game_id not in top_juegos_similares:
            top_juegos_similares.append(game_id)
    
    # Obtener los nombres de los juegos recomendados
    juegos_recomendados = []
    for game_id in top_juegos_similares:
        item_name = training[training['item_id'] == game_id]['item_name']
        if not item_name.empty:
            item_name = item_name.values[0]
            juegos_recomendados.append({'id del juego': game_id, 'Nombre': item_name})

    
    salida = {'Juegos recomendados a {}'.format(user_id): juegos_recomendados}
    return salida


Se prueba la funcion

recomendacion_usuario('RandomSounds')

Se guarda la informacion

training_1.to_csv('training_1.csv', index=False)
with gzip.open('cosine_sim.pkl.gz', 'wb') as f: pickle.dump(cosine_sim, f)
print(f'Se guardó el archivo')

cosine_sim.shape
